# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_jhoicas_Inventario-api_internal_application_dto.CreateSupplierRequest:
    properties:
      email:
        type: string
      lead_time_days:
        type: integer
      name:
        type: string
      nit:
        type: string
      payment_term_days:
        type: integer
      phone:
        type: string
    type: object
...
  github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  github_com_jhoicas_Inventario-api_internal_application_dto.SupplierListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.SupplierResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  github_com_jhoicas_Inventario-api_internal_application_dto.SupplierResponse:
    properties:
      company_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      lead_time_days:
        type: integer
      name:
        type: string
      nit:
        type: string
      payment_term_days:
        type: integer
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_jhoicas_Inventario-api_internal_application_dto.UpdateSupplierRequest:
    properties:
      email:
        type: string
      lead_time_days:
        type: integer
      name:
        type: string
      nit:
        type: string
      payment_term_days:
        type: integer
      phone:
        type: string
    type: object
...
host: api.naturerp.ludoia.com
info:
  contact: {}
  description: API para el sistema ERP.
  title: Tu API ERP
  version: "1.0"
paths:
...
  /api/suppliers:
    get:
      parameters:
      - description: Buscar por nombre o NIT
        in: query
        name: search
        type: string
      - default: 20
        description: Límite
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.SupplierListResponse'
      security:
      - Bearer: []
      summary: Listar proveedores
      tags:
      - suppliers
    post:
      consumes:
      - application/json
      parameters:
      - description: Datos del proveedor
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.CreateSupplierRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.SupplierResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Crear proveedor
      tags:
      - suppliers
  /api/suppliers/{id}:
    get:
      parameters:
      - description: ID del proveedor
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.SupplierResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Obtener proveedor por ID
      tags:
      - suppliers
    put:
      consumes:
      - application/json
      parameters:
      - description: ID del proveedor
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.UpdateSupplierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.SupplierResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Actualizar proveedor
      tags:
      - suppliers
  /api/suppliers/{id}/deactivate:
    put:
      description: Desactiva (soft delete) un proveedor asociado a la empresa autenticada
        (solo admin)
      parameters:
      - description: ID del proveedor
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Desactivar proveedor
      tags:
      - suppliers
...
securityDefinitions:
  Bearer:
    description: |-
      Escribe "Bearer " seguido de un espacio y tu token.
      Escribe "Bearer " seguido de un espacio y tu token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"